+++ hw24-webServer/src/main/java/ru/otus/WebServerWithBasicSecurity.java
@@ -5,22 +5,41 @@ import org.eclipse.jetty.security.HashLoginService;
 import org.eclipse.jetty.security.LoginService;
 import org.eclipse.jetty.util.resource.PathResourceFactory;
 import org.eclipse.jetty.util.resource.Resource;
+import org.hibernate.cfg.Configuration;
+import ru.otus.core.repository.DataTemplateHibernate;
+import ru.otus.core.repository.HibernateUtils;
+import ru.otus.core.sessionmanager.TransactionManagerHibernate;
+import ru.otus.crm.dbmigrations.MigrationsExecutorFlyway;
+import ru.otus.crm.model.Address;
+import ru.otus.crm.model.Client;
+import ru.otus.crm.model.Phone;
+import ru.otus.crm.service.DBServiceClient;
+import ru.otus.crm.service.DbServiceClientImpl;
 import ru.otus.crm.service.TemplateProcessor;
 import ru.otus.crm.service.TemplateProcessorImpl;
-import ru.otus.dao.ClientDao;
-import ru.otus.dao.HibernateClientDao;
 import ru.otus.helpers.FileSystemHelper;
-import ru.otus.server.UsersWebServer;
-import ru.otus.server.UsersWebServerWithBasicSecurity;
+import ru.otus.server.ClientsWebServer;
+import ru.otus.server.ClientsWebServerWithBasicSecurity;
 
 public class WebServerWithBasicSecurity {
     private static final int WEB_SERVER_PORT = 8080;
     private static final String TEMPLATES_DIR = "/templates/";
     private static final String HASH_LOGIN_SERVICE_CONFIG_NAME = "realm.properties";
     private static final String REALM_NAME = "AnyRealm";
+    public static final String HIBERNATE_CFG_FILE = "hibernate.cfg.xml";
 
     public static void main(String[] args) throws Exception {
-        ClientDao clientDao = new HibernateClientDao();
+        var configuration = new Configuration().configure(HIBERNATE_CFG_FILE);
+        var dbUrl = configuration.getProperty("hibernate.connection.url");
+        var dbUserName = configuration.getProperty("hibernate.connection.username");
+        var dbPassword = configuration.getProperty("hibernate.connection.password");
+
+        new MigrationsExecutorFlyway(dbUrl, dbUserName, dbPassword).executeMigrations();
+        var sessionFactory =
+                HibernateUtils.buildSessionFactory(configuration, Client.class, Address.class, Phone.class);
+        var transactionManager = new TransactionManagerHibernate(sessionFactory);
+        var clientTemplate = new DataTemplateHibernate<>(Client.class);
+        DBServiceClient dbServiceClient = new DbServiceClientImpl(transactionManager, clientTemplate);
         TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
 
         String hashLoginServiceConfigPath =
@@ -30,10 +49,10 @@ public class WebServerWithBasicSecurity {
 
         LoginService loginService = new HashLoginService(REALM_NAME, configResource);
 
-        UsersWebServer usersWebServer =
-                new UsersWebServerWithBasicSecurity(clientDao, WEB_SERVER_PORT, loginService, templateProcessor);
+        ClientsWebServer clientsWebServer = new ClientsWebServerWithBasicSecurity(
+                dbServiceClient, WEB_SERVER_PORT, loginService, templateProcessor);
 
-        usersWebServer.start();
-        usersWebServer.join();
+        clientsWebServer.start();
+        clientsWebServer.join();
     }
 }
