--- hw24-webServer/src/main/java/ru/otus/crm/model/User.java
@@ -1,32 +1,0 @@
-package ru.otus.crm.model;
-
-public class User {
-
-    private final long id;
-    private final String name;
-    private final String login;
-    private final String password;
-
-    public User(long id, String name, String login, String password) {
-        this.id = id;
-        this.name = name;
-        this.login = login;
-        this.password = password;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
