--- hw24-webServer/src/main/java/ru/otus/dao/ClientDao.java
@@ -1,11 +1,0 @@
-package ru.otus.dao;
-
-import java.util.List;
-import ru.otus.crm.model.Client;
-
-public interface ClientDao {
-
-    List<Client> findAll();
-
-    void add(Client client);
-}
