+++ hw24-webServer/src/main/java/ru/otus/server/ClientsWebServer.java
@@ -1,7 +1,7 @@
 package ru.otus.server;
 
 @SuppressWarnings({"squid:S112"})
-public interface UsersWebServer {
+public interface ClientsWebServer {
     void start() throws Exception;
 
     void join() throws Exception;
